requirements.txt
streamlit
pandas
numpy
scikit-learn
torch
transformers
nltk
google-generativeai
